refactored seccomp filters whitelist
Use one define on both aarch64 and x86_64.

For the syscall defines that are not available on musl+aarch64,
created a module which is aarch64 specific. The defines from the
module are exported on aarch64, while for x86_64 we use the libc
ones. This makes the code uniform on both platforms.

Additional syscalls need to be whitelisted on aarch64 as follows:
- SYS_rt_sigreturn
- SYS_io_submit

Signed-off-by: Andreea Florescu <fandree@amazon.com>