loader: add support for initrd
Add support for booting the micro VM with a user provided initrd file:
- For x86_64, the initrd is loaded at the end of the low memory,
and passed to the kernel in the bootparams table.
- For aarch64, the initrd is loaded in a memory segment just before the
FDT.  The initrd start and end addresses are specified in the FDT
"/chosen" node.

Users can specify an initrd file using the "initrd_path" property
of the "boot-source" the REST API.

This also adds integration tests to verify VMs boot with initrd and
without a rootfs. Categories of tests added:
- Functional: the VM shall boot and the "/" path should be mounted as
"rootfs" filesystem type.
- Performance: VM shall boot in less than 160ms.

Co-developed-by: Tim Deegan <tdeegan@amazon.com>
Signed-off-by: Tim Deegan <tdeegan@amazon.com>
Co-developed-by: Marco Vedovati <mvedovati@suse.com>
Signed-off-by: Marco Vedovati <mvedovati@suse.com>