Added EventManager trait
Introduced the EventManager trait with the following functions:
register, unregister, modify, and subscriber. Previously the
name `EventManager` was used for the implementation of the event
manager. This one is now renamed to `EpollManager` as it uses epoll
to handle events. The main drive for creating this trait is the need
to pass a reference to the event manager when calling process in
Subscriber. The only methods that are  safe to call are the ones
now defined in the trait.

Signed-off-by: Andreea Florescu <fandree@amazon.com>