some feedback + one lazy static for METRICS
I made some adjustments to the Metrics struct so that:
- we can use just one lazy static for metrics
- FirecrackerMetrics are not by default used, but anyone can
  override the application metrics with whatever they see fit

All feedback is written with TODO. Hopefully it is easy to spot.

P.S. The code doesn't build at the moment and basically every single
line that touches the METRICS instance needs to be changed from
METRICS.whatever_metric to METRICS.app_metrics.whatever_metric. If
you like the approach, we can make it build later.

Signed-off-by: Andreea Florescu <fandree@amazon.com>